wayne/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
sqlalchemy>=2.0.0
pdfplumber>=0.11.0
google-api-python-client>=2.140.0
google-auth-oauthlib>=1.2.1
google-auth>=2.34.0
python-dotenv>=1.0.1
aiofiles>=24.1.0
psutil>=6.0.0
pyperclip>=1.9.0
websockets>=12.0
openai-whisper
torch
numpy
soundfile
webrtcvad-wheels
rich>=13.8.0
prompt_toolkit>=3.0.47
ollama>=0.4.0
pydantic>=2.8.0
requests>=2.32.0
PyJWT[crypto]>=2.8.0
pvporcupine>=3.0.0
PyAudio>=0.2.14
SpeechRecognition>=3.10.0
pocketsphinx>=5.0.0
psycopg2-binary>=2.9.9
watchdog>=4.0.0
python-docx>=1.1.2
openpyxl>=3.1.5
python-pptx>=1.0.2
Pillow>=10.4.0
speedtest-cli>=2.1.3
pywin32>=306; sys_platform == "win32"
